pyproject.toml
[tool.poetry]
name = "fixmate-cli"
version = "0.0.55"
description = ""
authors = ["fixed.codes"]
license = "AGPL"
readme = "README.md"
packages = [
    { include = "fixmate", from = "." }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.8"
typing-extensions = "^4.11.0"
click = "~8.1.0"
openai = "^1.13.3"
tree-sitter-languages = "~1.10.2"
python-gitlab = "^4.4.0"
pygithub = "~2.1.1"
gitpython = "~3.1.40"
pydantic = "~2.8.2"
tiktoken = "~0.6.0"
libcst = "~1.1.0"
fixed-code2prompt = "0.2.0"
pyyaml = "^6.0.1"
packageurl-python = "~0.15.0"
semver = "~3.0.2"
requests = "~2.31.0"
chardet = "~5.2.0"
attrs = "~23.2.0"
google-ai-generativelanguage = "~0.6.7"
anthropic = "~0.31.2"
rich = "~13.7.1"
chevron = "~0.14.0"
scikit-learn = "^1.3.2"
# pinning transitive dependencies
tree-sitter = "~0.21.3"
numpy = [
    { version = "^1.26", python = "^3.12", optional = true},
    { version = "^1.24", python = "^3.8, <=3.11", optional = true}
]
pandas = [
    { version = "^2.1.4", python = "^3.12", optional = true},
    { version = "^2.0.3", python = "3.8, <=3.11", optional = true}
]
scipy = [
    { version = "^1.7.3", python = "^3.12", optional = true},
    { version = "^1.6.1", python = "3.8, <=3.11", optional = true}
]

# extras
# security
semgrep = {version = "^1.51.0", optional = true}
owasp-depscan = {version = "5.2.15", optional = true }

# rag
chromadb = {version = "~0.4.24", optional = true }
sentence-transformers = {version = "~2.7.0", optional = true }
# pinning transitive dependencies
torch = {version = "~2.2.2", optional = true }
orjson = {version = "~3.9.15", optional = true }

# notifications
slack-sdk = {version = "~3.28.0", optional = true }

[tool.poetry.extras]
security = ["semgrep", "owasp-depscan"]
rag = ["chromadb", "sentence-transformers", "torch", "orjson"]
notifications = ["slack-sdk"]
all = [
    "semgrep", "owasp-depscan",
    "chromadb", "sentence-transformers", "torch", "orjson",
    "slack-sdk",
]

[tool.poetry.group.dev.dependencies]
setuptools = "*"
poethepoet = { version = "^0.27.0", extras = ["poetry-plugin"] }
mypy = "^1.7.1"
types-requests = "~2.31.0"
black = "^23.12.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
pytest = "^8.1.1"
pytest-mock = "^3.8.0"
pytest-localserver = "^0.8.1"
datasets = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fixmate = 'fixmate.app:cli'

[tool.poe.poetry_hooks]
#pre_build = "mypy"
post_update = "lint"

[tool.poe.env]
PROJ_PATH.default = "fixmate"
TEST_PATH.default = "tests"

[tool.poe.tasks]
#mypy = "mypy ${PROJ_PATH}"
lint = [
    {cmd = "autoflake --recursive ${PROJ_PATH} ${TEST_PATH}"},
    {cmd = "isort ${PROJ_PATH} ${TEST_PATH}"},
    {cmd = "black ${PROJ_PATH} ${TEST_PATH}"}
]

[tool.black]
target-version = ['py310']
line-length = 120

[tool.isort]
profile = "black"

[tool.autoflake]
in-place = true
remove-all-unused-imports = true
expand-star-imports = true
ignore-init-module-imports = true
